Add a confirmation dialog to the delete button

+++ app/src/main/java/cz/dobris/zkousec/QuestionPackSetup.kt
@@ -1,5 +1,7 @@
 package cz.dobris.zkousec
 
+import android.app.AlertDialog
+import android.content.DialogInterface
 import android.os.Bundle
 import androidx.appcompat.app.AppCompatActivity
 import kotlinx.android.synthetic.main.activity_question_pack_setup2.*
@@ -17,7 +19,17 @@ class QuestionPackSetup : AppCompatActivity() {
             }
 
             DeleteButton.setOnClickListener {
-                Storage.deleteQFile(fileName, it.context)
+                AlertDialog.Builder(this)
+                    .setTitle("Delete file")
+                    .setMessage("Do you really want to delete the questions titled: " + fileName + "?")
+                    .setPositiveButton("Yes",
+                        DialogInterface.OnClickListener { dialog, which ->
+                            Storage.deleteQFile(fileName, it.context)
+                        })
+                    .setNegativeButton("No", DialogInterface.OnClickListener { dialog, which ->
+                    })
+                    .setIcon(android.R.drawable.ic_dialog_alert)
+                    .show()
             }
         }
     }
